app/src/requirements.txt
absl-py==0.7.0
alabaster==0.7.10
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
AnyQt==0.0.10
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
branca==0.3.1
breadability==0.1.20
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
cmake==3.12.0
colorama==0.3.9
commonmark==0.8.1
comtypes==1.1.4
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.3.2
defusedxml==0.5.0
dill==0.2.8.2
distributed==1.20.2
Django==2.1.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-model-utils==3.1.2
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
feature-selector===n-a
feedfinder2==0.0.4
feedparser==5.2.1
filelock==2.0.13
fileupload==0.1.5
Flask==0.12.2
Flask-Cors==3.0.3
folium==0.7.0
future==0.17.1
gast==0.2.2
gevent==1.2.2
glob2==0.6
google==2.0.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
graphviz==0.8.4
greenlet==0.4.12
grid==0.7.1
grpcio==1.16.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.12.1
hyperas==0.4.1
hyperopt==0.1.2
idna==2.7
imageio==2.2.0
imageio-ffmpeg==0.2.0
imagesize==0.7.1
ipykernel==5.1.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.13.3
jieba3k==0.35.1
Jinja2==2.10
joblib==0.13.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==16.0.0
keyrings.alt==3.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lightgbm==2.2.3
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.2
maze==1.3
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mlxtend==0.15.0.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
newspaper3k==0.2.8
nltk==3.2.5
nose==1.3.7
notebook==5.7.4
numba==0.36.2
numexpr==2.6.4
numpy==1.16.2
numpydoc==0.7.0
oauth2client==4.1.3
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
Orange3==3.17.0
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.3.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.3
py==1.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydot==1.4.1
pydotplus==2.0.2
PyDrive==1.3.1
pyflakes==1.6.0
Pygments==2.3.1
pylint==1.8.2
pymongo==3.7.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyqtgraph==0.10.0
pyreadline==2.1
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
python-louvain==0.11
python3-openid==3.1.0
pytz==2018.5
PyWavelets==0.5.2
pywin32==222
pywinpty==0.5
PyYAML==3.12
pyzmq==18.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.5.2
reportlab==3.5.9
requests==2.20.0
requests-file==1.4.3
requests-oauthlib==1.0.0
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.20.3
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
serverfiles==0.2.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
statsmodels==0.8.0
sumy==0.7.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.4
tinysegmenter==0.3
tldextract==2.2.1
toolz==0.9.0
tornado==4.5.3
tqdm==4.31.1
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
utils==0.9.0
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wget==3.2
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
zict==0.1.3
